observer: Handle reconnecting and being disconnected
We particularly need to keep track of the observed observations because
our protocol doesn't include explicit acknowledgements. We simply use
the observation getting cancelled as a signal it was processed. So we
keep track of all observations that we were told of, until we get a
message cancelling them (or they aren't included in the list we get from
SnoozeThis when starting up).