scanners/common: Protect the observableStream with a mutex
It wouldn't go wrong in practice, because there wouldn't be any
observables until you were connected so it could never try to notify of
completion until the observableStream was set at least once.

It was a data race reading it while we might overwrite it with a new
connection though.